mcp_genomic_data/requirements.txt
# MCP Genomic Data Server Requirements

# MCP SDK
mcp>=0.1.0

# Genomic data libraries
biopython>=1.81
pyfaidx>=0.7.0
cooler>=0.9.0
cooltools>=0.5.0

# Data access
requests>=2.31.0  # For API access (future)
aiohttp>=3.9.0   # Async HTTP (future)







